requirements.txt
pdfminer.six
python-docx
weasyprint
Jinja2
google-generativeai
python-dotenv
google-genai

